+++ cmd/get/catalogs/command.go
@@ -91,13 +91,14 @@ func New(config Config) (*cobra.Command, error) {
 	}
 
 	c := &cobra.Command{
-		Use:     name,
-		Short:   shortDescription,
-		Long:    longDescription,
-		Example: examples,
-		Aliases: []string{alias},
-		Args:    cobra.MaximumNArgs(1),
-		RunE:    r.Run,
+		Deprecated: "without replacement",
+		Use:        name,
+		Short:      shortDescription,
+		Long:       longDescription,
+		Example:    examples,
+		Aliases:    []string{alias},
+		Args:       cobra.MaximumNArgs(1),
+		RunE:       r.Run,
 		PreRunE: middleware.Compose(
 			renewtoken.Middleware(*config.ConfigFlags),
 		),