+++ cmd/gitops/add/app/command.go
@@ -99,11 +99,12 @@ func New(config Config) (*cobra.Command, error) {
 	}
 
 	c := &cobra.Command{
-		Use:     name,
-		Short:   shortDescription,
-		Long:    longDescription,
-		Example: examples,
-		RunE:    r.Run,
+		Deprecated: "currently without replacement",
+		Use:        name,
+		Short:      shortDescription,
+		Long:       longDescription,
+		Example:    examples,
+		RunE:       r.Run,
 	}
 
 	f.Init(c)