+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Added
 
 - Add locks to cache writing when doing `kubectl-gs login` to prevent race-condition between multiple kubectl calls.
+- Add extra logging when writing to cache.
 
 ## [4.11.0] - 2026-02-20